Move params defaults to storage.py && fix tests for new defaults
