+++ tests/test_bytes_array.py
@@ -22,6 +22,6 @@ def test_bytes_array(arr, gil):
 
 @pytest.mark.parametrize("data", [bytearray(7241), bytearray(7241) * 7])
 def test_bytearray(data):
-    cdata = blosc2.compress(data)
+    cdata = blosc2.compress(data, typesize=1)
     uncomp = blosc2.decompress(cdata)
     assert data == uncomp
